requirements.txt
firebase-admin==6.8.0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
